Fix ticket 362: display error for Blood DK Death Strike healing
The "player's healing-by-spell" submodule was intermittently hiding the bar for
the Death Strike heal of Blood DKs in segments where the DK cast Death Strike
while Dancing Rune Weapon was active. The DS heal from the player and rune
weapon were aliased in the bar display system, causing the bar for one or other
to disappear from the submodule display (but not from the tooltip). The DS
healing was still correctly accounted for and correctly displayed in other
modules (notably the total healing done by the DK).